backend/requirements.txt
Django>=5.2
psycopg2>=2.9.10
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
Faker>=13.1.0
